src/main/java/com/example/actionwithspring/BookController.java
package com.example.actionwithspring;


import com.example.actionwithspring.book.Book;
import com.example.actionwithspring.book.BookService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequiredArgsConstructor
@RequestMapping("/book")
public class BookController {
    private final BookService bookService;

    @GetMapping(consumes = MediaType.APPLICATION_JSON_VALUE,
            headers = {"Content-Type=application/json", "Accept=application/json"})
    public ResponseEntity getBooks(){
        return ResponseEntity.status(HttpStatus.OK).body(bookService.getBookList());
    }

    @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE,
            headers = {"Content-Type=application/json", "Accept=application/json"})
    public ResponseEntity addBook(@RequestBody Book book){
        bookService.addBook(book);
        return ResponseEntity.status(HttpStatus.OK).build();
    }
}
